api/router/enterprise_routes.go
package router

import (
	"net/http"
	"github.com/gorilla/mux"
	"github.com/emailimmunity/passwordimmunity/api/middleware"
	"github.com/emailimmunity/passwordimmunity/api/handlers"
	"github.com/emailimmunity/passwordimmunity/services/featureflag"
	"github.com/emailimmunity/passwordimmunity/services/licensing"
	"github.com/emailimmunity/passwordimmunity/services/payment"
)

// RegisterEnterpriseRoutes adds all enterprise-specific routes with license verification
func RegisterEnterpriseRoutes(
	r *mux.Router,
	featureManager *featureflag.FeatureManager,
	paymentService *payment.Service,
	enterpriseFeatureHandler *handlers.EnterpriseFeatureHandler,
) {
	// Enterprise middleware stack
	enterpriseMiddleware := []mux.MiddlewareFunc{
		middleware.RequireAuthentication,
		middleware.RequireEnterpriseLicense(featureManager.LicenseVerifier, ""),
	}

	// Enterprise routes
	enterprise := r.PathPrefix("/api/enterprise").Subrouter()
	for _, mw := range enterpriseMiddleware {
		enterprise.Use(mw)
	}

	// Feature management routes
	features := enterprise.PathPrefix("/features").Subrouter()
	features.HandleFunc("/activate", enterpriseFeatureHandler.ActivateFeature).Methods(http.MethodPost)
	features.HandleFunc("/deactivate", enterpriseFeatureHandler.DeactivateFeature).Methods(http.MethodPost)
	features.HandleFunc("/status", enterpriseFeatureHandler.GetActiveFeatures).Methods(http.MethodGet)
	features.HandleFunc("/bundles/activate", enterpriseFeatureHandler.ActivateBundle).Methods(http.MethodPost)
	features.HandleFunc("/bundles", handlers.HandleListBundles).Methods(http.MethodGet)
	features.HandleFunc("/available", handlers.HandleListAvailableFeatures).Methods(http.MethodGet)

	// Feature activation routes
	activation := enterprise.PathPrefix("/activation").Subrouter()
	activation.HandleFunc("/initiate", handlers.HandleFeatureActivation).Methods(http.MethodPost)
	activation.HandleFunc("/webhook", handlers.HandlePaymentWebhook).Methods(http.MethodPost)
	activation.HandleFunc("/status/{feature_id}", handlers.HandleFeatureStatus).Methods(http.MethodGet)

	// SSO routes
	sso := enterprise.PathPrefix("/sso").Subrouter()
	sso.Use(middleware.RequireEnterpriseLicense(featureManager.LicenseVerifier, licensing.FeatureAdvancedSSO))
	sso.Use(middleware.RequireFeature("advanced_sso"))
	sso.HandleFunc("/saml/metadata", handleSAMLMetadata).Methods(http.MethodGet)
	sso.HandleFunc("/saml/acs", handleSAMLACS).Methods(http.MethodPost)
	sso.HandleFunc("/oidc/config", handleOIDCConfig).Methods(http.MethodGet)

	// Custom roles routes
	roles := enterprise.PathPrefix("/roles").Subrouter()
	roles.Use(middleware.RequireEnterpriseLicense(featureManager.LicenseVerifier, licensing.FeatureCustomRoles))
	roles.Use(middleware.RequireFeature("custom_roles"))
	roles.HandleFunc("", handleListCustomRoles).Methods(http.MethodGet)
	roles.HandleFunc("", handleCreateCustomRole).Methods(http.MethodPost)
	roles.HandleFunc("/{id}", handleUpdateCustomRole).Methods(http.MethodPut)
	roles.HandleFunc("/{id}", handleDeleteCustomRole).Methods(http.MethodDelete)

	// Advanced reporting routes
	reporting := enterprise.PathPrefix("/reports").Subrouter()
	reporting.Use(middleware.RequireEnterpriseLicense(featureManager.LicenseVerifier, licensing.FeatureAdvancedReporting))
	reporting.Use(middleware.RequireFeature("advanced_reporting"))
	reporting.HandleFunc("/audit", handleAuditReport).Methods(http.MethodGet)
	reporting.HandleFunc("/usage", handleUsageReport).Methods(http.MethodGet)
	reporting.HandleFunc("/security", handleSecurityReport).Methods(http.MethodGet)

	// Multi-tenant management routes
	tenants := enterprise.PathPrefix("/tenants").Subrouter()
	tenants.Use(middleware.RequireEnterpriseLicense(featureManager.LicenseVerifier, licensing.FeatureMultiTenant))
	tenants.Use(middleware.RequireFeature("multi_tenant"))
	tenants.HandleFunc("", handleListTenants).Methods(http.MethodGet)
	tenants.HandleFunc("", handleCreateTenant).Methods(http.MethodPost)
	tenants.HandleFunc("/{id}", handleUpdateTenant).Methods(http.MethodPut)
	tenants.HandleFunc("/{id}", handleDeleteTenant).Methods(http.MethodDelete)

	// Security bundle routes
	security := enterprise.PathPrefix("/security").Subrouter()
	security.Use(middleware.RequireBundle("security"))
	security.HandleFunc("/policies", handleSecurityPolicies).Methods(http.MethodGet)
	security.HandleFunc("/audit", handleSecurityAudit).Methods(http.MethodGet)
}

// Handler stubs - to be implemented in separate files
func handleSAMLMetadata(w http.ResponseWriter, r *http.Request)  {}
func handleSAMLACS(w http.ResponseWriter, r *http.Request)       {}
func handleOIDCConfig(w http.ResponseWriter, r *http.Request)    {}
func handleListCustomRoles(w http.ResponseWriter, r *http.Request)    {}
func handleCreateCustomRole(w http.ResponseWriter, r *http.Request)   {}
func handleUpdateCustomRole(w http.ResponseWriter, r *http.Request)   {}
func handleDeleteCustomRole(w http.ResponseWriter, r *http.Request)   {}
func handleAuditReport(w http.ResponseWriter, r *http.Request)       {}
func handleUsageReport(w http.ResponseWriter, r *http.Request)       {}
func handleSecurityReport(w http.ResponseWriter, r *http.Request)    {}
func handleListTenants(w http.ResponseWriter, r *http.Request)       {}
func handleCreateTenant(w http.ResponseWriter, r *http.Request)      {}
func handleUpdateTenant(w http.ResponseWriter, r *http.Request)      {}
func handleDeleteTenant(w http.ResponseWriter, r *http.Request)      {}

// Security bundle handlers
func handleSecurityPolicies(w http.ResponseWriter, r *http.Request) {
	// TODO: Implement security policies handler
	w.WriteHeader(http.StatusNotImplemented)
}

func handleSecurityAudit(w http.ResponseWriter, r *http.Request) {
	// TODO: Implement security audit handler
	w.WriteHeader(http.StatusNotImplemented)
}
